refactor: Coluna unidade da tabela produto modificada de BIGINT para INTEGER.

+++ src/main/resources/db/migration/V1__create_table_produto.sql
@@ -2,6 +2,6 @@ CREATE TABLE produto(
     id BIGSERIAL PRIMARY KEY,
     codigo_produto VARCHAR(50) UNIQUE NOT NULL,
     nome VARCHAR(100) NOT NULL,
-    unidade BIGINT NOT NULL,
+    unidade INTEGER NOT NULL,
     validade DATE NOT NULL
 );
